requirements.txt
fastapi
uvicorn
azure-storage-file-datalake
python-multipart